Optimization and clean-up: v0.0.9
This update gets rid of a few classes that are too specific to be generalized, such as the TimeCycle class. This optimizes several classes, both in the way of that they are easier to explain and have easier ways to execute methods. The biggest change was the implementation of AngleF, which automatically handles degree to radian conversion. Lastly, Encoding/Raw was restored to pave the way for GifDecoder, which would allow you to load GIF files and continue editing them.