Optimize direct Bitmap: v0.0.10
This small patch optimizes the GetNonEmptyWidth method to use direct pixels to count instead of the hefty GetPixel method. The ability to force a set of colors to an image has also been introduced, alongside getting a Grid<Color> GetPixels() method for a Bitmap, which allows you to deconstruct a Bitmap to a grid of color values that can easily be read. The AngleF class was also changed to support extra operator methods between float values to make the class extremely easy to utilize.