Update timeline methods and organization: v0.0.12
This update was quite heavy to implement. The first steps were to rename and organize everything accordingly. I removed more classes that felt redundant to use, and I implemented functional mechanics in relation to the TimelineAnimator. The animator now correctly utilizes keyframes in order to create an animation that transitions from one position to the next in a clean manner. More documentation was added for some classes, and all rotation/transform methods now work as intended.