Quick refactoring: v0.1.1
This update adjusts and cleans up a couple of methods, while also introducing a new concept to possible replace the Grid<T> class in terms of image manipulation, called BitmapSource (unimplemented). This will attempt to use the methods specified in ImageEditor and ImageHelper to edit images in a much faster way than the Grid<T> class.